Can add arbitrary tags to break on
Needed to add a prompt for users to be able to specify any tag to break
on. This is the same functionality from the original M Debugger.

+++ gov.va.med.iss.debug.ui/src/gov/va/mumps/debug/ui/model/MLineBreakpointAdapter.java
@@ -28,7 +28,8 @@ public class MLineBreakpointAdapter implements IToggleBreakpointsTarget {
 			IBreakpoint[] breakpoints = DebugPlugin.getDefault().getBreakpointManager().getBreakpoints(MDebugConstants.M_DEBUG_MODEL);
 			for (int i = 0; i < breakpoints.length; i++) {
 				IBreakpoint breakpoint = breakpoints[i];
-				if (resource.equals(breakpoint.getMarker().getResource())) {
+				if (resource.equals(breakpoint.getMarker().getResource()) &&
+						breakpoint instanceof ILineBreakpoint) {
 					if (((ILineBreakpoint)breakpoint).getLineNumber() == (lineNumber + 1)) {
 						// remove
 						breakpoint.delete();
